+++ templates_c/ba88594facaf868cff2f0d86ade1fd1c639a1500_0.file.libros.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 3.1.39, created on 2021-11-28 04:51:58
+/* Smarty version 3.1.39, created on 2021-11-28 20:51:26
   from 'C:\xampp\htdocs\Tpe 2 web\TPEspecial\templates\libros.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '3.1.39',
-  'unifunc' => 'content_61a2fcde640f38_25354323',
+  'unifunc' => 'content_61a3ddbecd3050_69224184',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     'ba88594facaf868cff2f0d86ade1fd1c639a1500' => 
     array (
       0 => 'C:\\xampp\\htdocs\\Tpe 2 web\\TPEspecial\\templates\\libros.tpl',
-      1 => 1638070732,
+      1 => 1638127630,
       2 => 'file',
     ),
   ),
@@ -22,7 +22,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:templates/footer.tpl' => 1,
   ),
 ),false)) {
-function content_61a2fcde640f38_25354323 (Smarty_Internal_Template $_smarty_tpl) {
+function content_61a3ddbecd3050_69224184 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_checkPlugins(array(0=>array('file'=>'C:\\xampp\\htdocs\\Tpe2web\\TPEspecial\\libs\\smarty-3.1.39\\libs\\plugins\\modifier.truncate.php','function'=>'smarty_modifier_truncate',),));
 $_smarty_tpl->_subTemplateRender("file:templates/header.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
@@ -47,13 +47,15 @@ $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
     </p>
     <p> Login:
       <a class="btn btn-secondary mt-2 mb-2" href="login">Login</a>
-      </p>
+    </p>
     <p> Logout:
         <a class="btn btn-secondary mt-2 mb-2" href="logout">Logout</a>
     </p>
+    <?php if ($_smarty_tpl->tpl_vars['rolAndId']->value[1] == 1) {?>
         <p> admin:
-        <a class="btn btn-secondary mt-2 mb-2" href="adminView">Admin</a>
-    </p>
+            <a class="btn btn-secondary mt-2 mb-2" href="adminView">Admin</a>
+        </p>
+    <?php }?>
 </nav>
 
 <table  class="table table-dark table-hover">
@@ -62,7 +64,9 @@ $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
             <th>Genero</th>
             <th>Descripcion</th>
             <th><hr/></th>
-            <th>Eliminar</th>
+            <?php if ($_smarty_tpl->tpl_vars['rolAndId']->value[1] == 1) {?>
+                <th>Eliminar</th>
+            <?php }?>
         </thead>
         <tbody>
         <?php
@@ -80,8 +84,10 @@ $_smarty_tpl->tpl_vars['book']->do_else = false;
 </td>
                 <td> <a class="btn btn-info" href="viewDescripcion/<?php echo $_smarty_tpl->tpl_vars['book']->value->id_libros;?>
 ">Leer mas...</a> </td>
-                <td> <a class="btn btn-danger" href="deleteBook/<?php echo $_smarty_tpl->tpl_vars['book']->value->id_libros;?>
+                <?php if ($_smarty_tpl->tpl_vars['rolAndId']->value[1] == 1) {?>
+                    <td> <a class="btn btn-danger" href="deleteBook/<?php echo $_smarty_tpl->tpl_vars['book']->value->id_libros;?>
 ">Eliminar</a> </td>
+                <?php }?>
             </tr>
         <?php
 }
@@ -89,48 +95,51 @@ $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
     </tbody>
 </table>
 
-<form action="agregarLibro" method="POST">
-         <label>TItulo: </label><input type="text" name="titulo">
-        <label>Genero: </label> <input type="text" name="genero">
-        <label>Descripcion: </label> <textarea name="texto"cols="30" rows="1"></textarea>
-        <label>Numero del autor</label>
-    <select name="select">
-        <?php
+<?php if ($_smarty_tpl->tpl_vars['rolAndId']->value[1] == 1) {?>
+
+    <form action="agregarLibro" method="POST">
+            <label>TItulo: </label><input type="text" name="titulo">
+            <label>Genero: </label> <input type="text" name="genero">
+            <label>Descripcion: </label> <textarea name="texto"cols="30" rows="1"></textarea>
+            <label>Numero del autor</label>
+        <select name="select">
+            <?php
 $_from = $_smarty_tpl->smarty->ext->_foreach->init($_smarty_tpl, $_smarty_tpl->tpl_vars['books']->value, 'book');
 $_smarty_tpl->tpl_vars['book']->do_else = true;
 if ($_from !== null) foreach ($_from as $_smarty_tpl->tpl_vars['book']->value) {
 $_smarty_tpl->tpl_vars['book']->do_else = false;
 ?>    
-            <option value="<?php echo $_smarty_tpl->tpl_vars['book']->value->fk_Id_autor;?>
+                <option value="<?php echo $_smarty_tpl->tpl_vars['book']->value->fk_Id_autor;?>
 "><?php echo $_smarty_tpl->tpl_vars['book']->value->Id_autor;?>
 </option>
-        <?php
+            <?php
 }
 $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
-    </select>
-        <input class="btn btn-success" type="submit" value="Enviar">
-</form>
+        </select>
+            <input class="btn btn-success" type="submit" value="Enviar">
+    </form>
 
-<form action="updateBook" method="POST">
-         <label>TItulo: </label><input type="text" name="titulo">
-        <label>Genero: </label> <input type="text" name="genero">
-        <label>Descripcion: </label> <textarea name="texto"cols="30" rows="1"></textarea>
-    <select name="select">
-        <?php
+    <form action="updateBook" method="POST">
+            <label>TItulo: </label><input type="text" name="titulo">
+            <label>Genero: </label> <input type="text" name="genero">
+            <label>Descripcion: </label> <textarea name="texto"cols="30" rows="1"></textarea>
+        <select name="select">
+            <?php
 $_from = $_smarty_tpl->smarty->ext->_foreach->init($_smarty_tpl, $_smarty_tpl->tpl_vars['books']->value, 'book');
 $_smarty_tpl->tpl_vars['book']->do_else = true;
 if ($_from !== null) foreach ($_from as $_smarty_tpl->tpl_vars['book']->value) {
 $_smarty_tpl->tpl_vars['book']->do_else = false;
 ?>    
-            <option value="<?php echo $_smarty_tpl->tpl_vars['book']->value->id_libros;?>
+                <option value="<?php echo $_smarty_tpl->tpl_vars['book']->value->id_libros;?>
 "><?php echo $_smarty_tpl->tpl_vars['book']->value->Titulo;?>
 </option>
-        <?php
+            <?php
 }
 $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
-    </select>
-        <input class="btn btn-success" type="submit" value="Modificar">
-</form>
+        </select>
+            <input class="btn btn-success" type="submit" value="Modificar">
+    </form>
+<?php }?>
 
 <?php $_smarty_tpl->_subTemplateRender("file:templates/footer.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 }
